+++ app/src/main/java/com/pavesid/androidacademy/data/movies/JsonMovie.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class JsonMovie(
     @SerialName("id")
-    val id: Int,
+    val id: Long,
     @SerialName("title")
     val title: String,
     @SerialName("overview")
@@ -16,7 +16,7 @@ data class JsonMovie(
     @SerialName("backdrop_path")
     val backdropPicture: String?,
     @SerialName("genre_ids")
-    val genreIds: List<Int>,
+    val genreIds: List<Long>,
     @SerialName("vote_average")
     val ratings: Float,
     @SerialName("vote_count")
